add basic setup.py file
fix version

it wasn't so simple...

undo lib rename hack
