+++ tachyon/core/textutils.py
@@ -16,7 +16,7 @@
 # Place, Suite 330, Boston, MA  02111-1307  USA
 #
 
-from core import database, conf
+from . import database, conf
 from datetime import datetime
