+++ tachyon/plugins/host/HostProcessor.py
@@ -15,7 +15,7 @@
 # this program; if not, write to the Free Software Foundation, Inc., 59 Temple
 # Place, Suite 330, Boston, MA  02111-1307  USA
 #
-from core import conf, textutils, database
+from ...core import conf, textutils, database
 
 def execute():
     """ This plugin process the hostname to generate host and filenames relatives to it """
