+++ tachyon/plugins/host/SitemapXML.py
@@ -17,8 +17,8 @@
 #
 
 import re
-from core import conf, textutils, database
-from core.fetcher import Fetcher
+from ...core import conf, textutils, database
+from ...core.fetcher import Fetcher
 try:
     from urlparse import urljoin
     from urlparse import urlparse
